模型管理代码规范调整

+++ qmodel-module-model/qmodel-module-model-biz/src/main/java/tech/qiantong/qmodel/module/model/controller/admin/model/vo/ModelSaveReqVO.java
@@ -104,7 +104,7 @@ public class ModelSaveReqVO extends BaseEntity {
     private String source;
 
     @Schema(description = "上传状态", example = "")
-    private Long uploadStatus;
+    private Integer uploadStatus;
 
     @Schema(description = "上传时间", example = "")
     @JsonFormat(pattern = "yyyy-dd HH:mm:ss")
